Fix: mcu firmware v2.5: Bootsel IO pins was changed by mistake
The bootsel pins was changed to input mode by mistake in
som_reset_control(), in other words, the bootsel was controlled only
by hardware dip switch and not by the software.
This is not expected behaviour when software controlled bootsel mode
is configured.
Deleted the set_bootsel() call in som_reset_control()

Signed-off-by: linmin <linmin@eswincomputing.com>
Signed-off-by: Darshan Jasanee <darshan.jasanee@einfochips.com>